+++ crates/types/src/error.rs
@@ -2,12 +2,10 @@ use std::error::Error as StdError;
 use std::ffi::{c_char, CStr, CString};
 use std::fmt;
 
-use thiserror::Error as ThisError;
-
 // https://github.com/neovim/neovim/blob/v0.9.0/src/nvim/api/private/defs.h#L64-L67
 //
 /// Binding to the error type used by Neovim.
-#[derive(Clone, ThisError, Eq, PartialEq, Hash)]
+#[derive(Clone, Eq, PartialEq, Hash)]
 #[repr(C)]
 pub struct Error {
     r#type: ErrorType,
@@ -16,6 +14,7 @@ pub struct Error {
 
 unsafe impl Send for Error {}
 unsafe impl Sync for Error {}
+impl StdError for Error {}
 
 // https://github.com/neovim/neovim/blob/v0.9.0/src/nvim/api/private/defs.h#L27-L31
 #[derive(Copy, Clone, Eq, PartialEq, Hash)]