+++ tests/src/api/global.rs
@@ -1,5 +1,9 @@
+use std::sync::Arc;
+
 use all_asserts::*;
 use nvim_oxi::api::{self, opts::*, types::*, Buffer, Window};
+use nvim_oxi::mlua::{Error as LuaError, IntoLuaMulti, Lua, Table};
+use nvim_oxi::{Dictionary, Object};
 
 #[nvim_oxi::test]
 fn chan_send_fail() {
@@ -177,6 +181,46 @@ fn list_wins() {
 }
 
 #[nvim_oxi::test]
+fn notify() {
+    let opts = Dictionary::new();
+    let ret = api::notify("", LogLevel::Error, &opts).unwrap();
+    assert_eq!(ret, Object::nil());
+}
+
+// Fails on 0.9.5 on macOS and Windows. Not sure why.
+#[nvim_oxi::test]
+#[cfg_attr(not(any(target_os = "linux", feature = "neovim-0-10")), ignore)]
+fn notify_custom() {
+    let message = "Notifier was called!";
+
+    // Set up a custom notification provider.
+    set_notification_provider(move |lua, _msg, _level, _opts| {
+        lua.create_string(message)
+    });
+
+    let opts = Dictionary::new();
+    let ret = api::notify("", LogLevel::Error, &opts).unwrap();
+    assert_eq!(ret, message.into());
+}
+
+// Fails on 0.9.5 on macOS and Windows. Not sure why.
+#[nvim_oxi::test]
+#[cfg_attr(not(any(target_os = "linux", feature = "neovim-0-10")), ignore)]
+fn notify_custom_err() {
+    #[derive(Debug, thiserror::Error)]
+    #[error("")]
+    struct CustomError;
+
+    // Set up a custom notification provider.
+    set_notification_provider(move |_lua, _msg, _level, _opts| {
+        Err::<(), _>(LuaError::ExternalError(Arc::new(CustomError)))
+    });
+
+    let opts = Dictionary::new();
+    let _err = api::notify("", LogLevel::Error, &opts).unwrap_err();
+}
+
+#[nvim_oxi::test]
 fn set_get_del_current_line() {
     let res = api::set_current_line("foo");
     assert_eq!(Ok(()), res);
@@ -272,4 +316,19 @@ fn hex_to_dec(hex_color: &str) -> u32 {
     assert!(hex_color[1..].chars().all(|c| c.is_ascii_digit()
         || ('a'..='f').contains(&c.to_ascii_lowercase())));
     u32::from_str_radix(&hex_color[1..], 16).unwrap()
+}
+
+fn set_notification_provider<P, R>(mut provider: P)
+where
+    P: FnMut(&Lua, String, u32, Table) -> Result<R, LuaError> + 'static,
+    R: IntoLuaMulti,
+{
+    let lua = nvim_oxi::mlua::lua();
+    let vim = lua.globals().get::<Table>("vim").unwrap();
+    let notify = lua
+        .create_function_mut(move |lua, (msg, level, opts)| {
+            provider(lua, msg, level, opts)
+        })
+        .unwrap();
+    vim.set("notify", notify).unwrap();
 }